Split commands like in split and str_ but do not register?
