+++ crates/nu-cli/tests/commands/average.rs
@@ -7,11 +7,11 @@ fn can_average_numbers() {
         r#"
              open sgml_description.json
              | get glossary.GlossDiv.GlossList.GlossEntry.Sections
-             | average
+             | math average
              | echo $it
          "#
     ));
-
+    println!("{:?}", actual.err);
     assert_eq!(actual.out, "101.5")
 }
 
@@ -19,7 +19,7 @@ fn can_average_numbers() {
 fn can_average_bytes() {
     let actual = nu!(
         cwd: "tests/fixtures/formats",
-        "ls | sort-by name | skip 1 | first 2 | get size | average | format \"{$it}\" | echo $it"
+        "ls | sort-by name | skip 1 | first 2 | get size | math average | format \"{$it}\" | echo $it"
     );
 
     assert_eq!(actual.out, "1.6 KB");
